# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.AddCartItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
      user_id:
        type: string
    type: object
  handlers.CheckoutRequest:
    properties:
      campaign_ids:
        items:
          type: string
        type: array
      point_used:
        type: integer
      user_id:
        type: string
    type: object
  handlers.CheckoutResponse:
    properties:
      message:
        type: string
      total_price:
        type: number
    type: object
  handlers.DeleteCartItemRequest:
    properties:
      product_id:
        type: string
      user_id:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.GuestLoginResponse:
    properties:
      Message:
        type: string
      id:
        type: string
      point:
        type: integer
    type: object
  handlers.RealignCategoryRequest:
    properties:
      category_id:
        type: string
      rank:
        type: integer
    type: object
  handlers.RegisterCampaign:
    properties:
      campaign_category_id:
        type: string
      description:
        type: string
      discount_type:
        type: string
      discount_value:
        type: number
      every:
        type: number
      is_active:
        type: boolean
      limit:
        type: number
      list_product_category_id:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
  handlers.RegisterCampaignCategory:
    properties:
      description:
        type: string
      name:
        type: string
      rank:
        type: integer
    type: object
  handlers.RegisterProduct:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      product_category_id:
        type: string
    type: object
  handlers.RegisterProductCategory:
    properties:
      name:
        type: string
    type: object
  models.Campaign:
    properties:
      campaign_category_id:
        type: string
      description:
        type: string
      discount_type:
        description: '"percent", "fixed", "point", "spendAndSave"'
        type: string
      discount_value:
        type: number
      every:
        type: number
      id:
        type: string
      is_active:
        type: boolean
      limit:
        type: number
      name:
        type: string
      product_categories:
        items:
          $ref: '#/definitions/models.ProductCategory'
        type: array
    type: object
  models.CampaignsCategories:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      rank:
        type: integer
    type: object
  models.CartItem:
    properties:
      id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      user_id:
        type: string
    type: object
  models.Product:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      product_category_id:
        type: string
      product_category_name:
        type: string
    type: object
  models.ProductCategory:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: API endpoints for registering users.
  title: Ecom backend api
  version: "1.0"
paths:
  /campaign-categories:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CampaignsCategories'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get all campaign categories
      tags:
      - Campaigns
    post:
      consumes:
      - application/json
      parameters:
      - description: Category payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterCampaignCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.RegisterCampaignCategory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new campaign category
      tags:
      - Campaigns
  /campaign-categories/realign:
    patch:
      consumes:
      - application/json
      description: Update ranks for multiple campaign categories
      parameters:
      - description: List of categories with new ranks
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/handlers.RealignCategoryRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Realign campaign category ranks
      tags:
      - Campaigns
  /campaigns:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Campaign'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get all campaigns
      tags:
      - Campaigns
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign payload
        in: body
        name: campaign
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterCampaign'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.RegisterCampaign'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new campaign
      tags:
      - Campaigns
  /campaigns/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a campaign by ID (set is_active to false)
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Soft delete a campaign
      tags:
      - Campaigns
  /campaigns/{id}/activate:
    patch:
      consumes:
      - application/json
      description: Activate a campaign by ID (set is_active to true)
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Activate a campaign
      tags:
      - Campaigns
  /cart:
    delete:
      consumes:
      - application/json
      description: Removes a product from the user's cart
      parameters:
      - description: Delete Cart Item payload
        in: body
        name: cartItem
        required: true
        schema:
          $ref: '#/definitions/handlers.DeleteCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Remove item from cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Adds a product to the user's cart or updates quantity if it already
        exists
      parameters:
      - description: Cart Item payload
        in: body
        name: cartItem
        required: true
        schema:
          $ref: '#/definitions/handlers.AddCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CartItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Add item to cart
      tags:
      - Cart
  /cart/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve all items in a user's cart with product details
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get cart items for a user
      tags:
      - Cart
  /checkout:
    post:
      consumes:
      - application/json
      description: Calculate total price, apply campaigns, and store transaction history
      parameters:
      - description: Checkout payload
        in: body
        name: checkout
        required: true
        schema:
          $ref: '#/definitions/handlers.CheckoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CheckoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Checkout cart items
      tags:
      - Checkout
  /guestregister:
    post:
      consumes:
      - application/json
      description: Creates a new guest user account.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.GuestLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register a guest user
      tags:
      - Auth
  /product-categories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProductCategory'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get all product categories
      tags:
      - Products
    post:
      consumes:
      - application/json
      parameters:
      - description: Category payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterProductCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductCategory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new product category
      tags:
      - Products
  /product-categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete a product category
      tags:
      - Products
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Adds a product with category and optional description.
      parameters:
      - description: Product payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new product
      tags:
      - Products
swagger: "2.0"